cast to int

+++ train_gpt2.py
@@ -193,7 +193,7 @@ class MLP(nn.Module):
 
     def __init__(self, config):
         super().__init__()
-        hidden_dim = config.n_embd * 43/16 # 11008/4096
+        hidden_dim = int(config.n_embd * 43/16) # 11008/4096
         self.c_fc    = CastedLinear(config.n_embd, hidden_dim, bias=False)
         self.c_proj  = CastedLinear(hidden_dim, config.n_embd, bias=False)
         self.c_proj.weight.data.zero_() # zero init suggested by @Grad62304977
